Huge update

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -11,8 +11,12 @@ class CreateUsersTable extends Migration
      *
      * @return void
      */
+
     public function up()
     {
+        if (Schema::hasTable('users')) {
+            //
+        } else {
         Schema::create('users', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
@@ -36,6 +40,7 @@ class CreateUsersTable extends Migration
             $table->rememberToken();
             $table->timestamps();
         });
+       }
     }
 
     /**
@@ -45,6 +50,10 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('users');
+        if (Schema::hasTable('users')) {
+            //
+        } else {
+            Schema::dropIfExists('users');
+        }
     }
 }
